Consolidate all color functions into color-utils.js
