+++ packages/block-library/src/cover/shared.js
@@ -23,7 +23,6 @@ export const COVER_MAX_HEIGHT = 1000;
 export const COVER_DEFAULT_HEIGHT = 300;
 export const DEFAULT_FOCAL_POINT = { x: 0.5, y: 0.5 };
 export const ALLOWED_MEDIA_TYPES = [ 'image', 'video' ];
-export const DEFAULT_AVERAGE_COLOR = '#FFFFFF';
 
 export function mediaPosition( { x, y } = DEFAULT_FOCAL_POINT ) {
 	return `${ Math.round( x * 100 ) }% ${ Math.round( y * 100 ) }%`;
